email_automation/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
WTForms==3.0.0
SQLAlchemy==1.4.23
apscheduler==3.9.1
pandas==1.3.3
email-validator==1.1.3
python-dotenv==0.19.1
bcrypt==3.2.0
gunicorn==20.1.0
pytz==2022.1
imaplib2==3.6
Flask-SocketIO==5.3.6
eventlet==0.33.3
gevent==22.10.2
gevent-websocket==0.10.1 